Update `usb` package to latest
This likely takes away error reporting to some level. I haven't done a deep dive into whether that top-level error event has an analog in the new version of the lib.
The main win here is that the use of `claim` allowed the tool to find my odrive (old mac, latest os)